[Refactor/#92] 공통 컴포넌트 core로 이동
