+++ app/src/main/java/com/swyp/firsttodo/core/common/component/HaebomDeleteDialog.kt
@@ -1,4 +1,4 @@
-package com.swyp.firsttodo.presentation.common.component
+package com.swyp.firsttodo.core.common.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -27,38 +27,9 @@ import androidx.compose.ui.unit.dp
 import com.swyp.firsttodo.core.base.Async
 import com.swyp.firsttodo.core.common.extension.noRippleClickable
 import com.swyp.firsttodo.core.common.extension.skeleton
+import com.swyp.firsttodo.core.common.type.DeleteDialogType
 import com.swyp.firsttodo.core.designsystem.component.HaebomBasicDialog
 import com.swyp.firsttodo.core.designsystem.theme.HaebomTheme
-
-sealed class DeleteDialogType(
-    val title: String,
-    val description: String,
-) {
-    data object Todo : DeleteDialogType(
-        title = "선택한 할 일을 삭제할까요?",
-        description = "입력한 할 일이 사라져요!",
-    )
-
-    data object Schedule : DeleteDialogType(
-        title = "선택한 일정을 삭제할까요?",
-        description = "입력한 일정이 사라져요!",
-    )
-
-    data object Habit : DeleteDialogType(
-        title = "선택한 습관을 삭제할까요?",
-        description = "입력한 습관이 사라져요!",
-    )
-
-    data object FailedHabit : DeleteDialogType(
-        title = "실패한 습관을 삭제할까요?",
-        description = "실패한 습관이 사라져요!",
-    )
-
-    class Disconnect(nickname: String) : DeleteDialogType(
-        title = "${nickname}님과의 연결을 끊을까요?",
-        description = "연동이 끊기면 ${nickname}님의\n활동을 볼 수 없습니다.",
-    )
-}
 
 @Composable
 fun HaebomDeleteDialog(
