+++ app/src/main/java/com/swyp/firsttodo/core/common/type/TaskItemPopupType.kt
@@ -1,0 +1,6 @@
+package com.swyp.firsttodo.core.common.type
+
+enum class TaskItemPopupType(val text: String) {
+    EDIT(text = "수정하기"),
+    RETRY(text = "재도전하기"),
+}
