+++ app/src/main/java/com/swyp/firsttodo/presentation/habit/component/HabitList.kt
@@ -28,11 +28,11 @@ import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import androidx.compose.ui.unit.dp
 import com.swyp.firsttodo.R
+import com.swyp.firsttodo.core.common.component.TaskItemPopup
 import com.swyp.firsttodo.core.common.extension.noRippleClickable
 import com.swyp.firsttodo.core.designsystem.theme.HaebomTheme
 import com.swyp.firsttodo.domain.model.habit.HabitDuration
 import com.swyp.firsttodo.domain.model.habit.HabitModel
-import com.swyp.firsttodo.presentation.common.component.task.TaskItemPopup
 
 enum class HabitListType {
     CHILD,
