+++ app/src/main/java/com/swyp/firsttodo/presentation/habit/component/HabitRetryList.kt
@@ -42,17 +42,17 @@ import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import androidx.compose.ui.unit.dp
 import com.swyp.firsttodo.R
 import com.swyp.firsttodo.core.base.Async
+import com.swyp.firsttodo.core.common.component.TUTORIAL_MIN_SPACE_DP
+import com.swyp.firsttodo.core.common.component.TUTORIAL_SCROLL_AMOUNT_DP
+import com.swyp.firsttodo.core.common.component.TaskItemPopup
+import com.swyp.firsttodo.core.common.component.TutorialOverlay
 import com.swyp.firsttodo.core.common.extension.getDataOrNull
 import com.swyp.firsttodo.core.common.extension.noRippleClickable
+import com.swyp.firsttodo.core.common.type.TaskItemPopupType
+import com.swyp.firsttodo.core.common.type.TutorialType
 import com.swyp.firsttodo.core.designsystem.theme.HaebomTheme
 import com.swyp.firsttodo.domain.model.habit.HabitDuration
 import com.swyp.firsttodo.domain.model.habit.HabitModel
-import com.swyp.firsttodo.presentation.common.component.TUTORIAL_MIN_SPACE_DP
-import com.swyp.firsttodo.presentation.common.component.TUTORIAL_SCROLL_AMOUNT_DP
-import com.swyp.firsttodo.presentation.common.component.TutorialOverlay
-import com.swyp.firsttodo.presentation.common.component.TutorialType
-import com.swyp.firsttodo.presentation.common.component.task.TaskItemPopup
-import com.swyp.firsttodo.presentation.common.component.task.TaskItemPopupType
 import kotlinx.coroutines.launch
 
 @Composable
