+++ app/src/main/java/com/swyp/firsttodo/presentation/hamburger/share/ShareScreen.kt
@@ -17,12 +17,12 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.swyp.firsttodo.core.base.Async
+import com.swyp.firsttodo.core.common.component.HaebomDeleteDialog
+import com.swyp.firsttodo.core.common.component.HaebomTopBar
+import com.swyp.firsttodo.core.common.type.DeleteDialogType
 import com.swyp.firsttodo.core.common.util.HandleSideEffects
 import com.swyp.firsttodo.core.designsystem.theme.HaebomTheme
 import com.swyp.firsttodo.domain.model.family.ConnectedFamilyModel
-import com.swyp.firsttodo.presentation.common.component.DeleteDialogType
-import com.swyp.firsttodo.presentation.common.component.HaebomDeleteDialog
-import com.swyp.firsttodo.presentation.common.component.HaebomTopBar
 import com.swyp.firsttodo.presentation.hamburger.share.component.CodeInputSection
 import com.swyp.firsttodo.presentation.hamburger.share.component.ConnectedFamilySection
 import com.swyp.firsttodo.presentation.hamburger.share.component.MyCodeSection
