+++ app/src/main/java/com/swyp/firsttodo/presentation/onboarding/component/ProfileView.kt
@@ -14,8 +14,8 @@ import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.swyp.firsttodo.core.common.component.HaebomMultiLineTextField
 import com.swyp.firsttodo.core.designsystem.theme.HaebomTheme
-import com.swyp.firsttodo.presentation.common.component.HaebomMultiLineTextField
 
 @Composable
 fun ProfileView(
