+++ app/src/main/java/com/swyp/firsttodo/presentation/reward/detail/RewardDetailScreen.kt
@@ -20,13 +20,13 @@ import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import androidx.compose.ui.unit.dp
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import com.swyp.firsttodo.core.common.component.HaebomLargeButton
+import com.swyp.firsttodo.core.common.component.HaebomMultiLineTextField
+import com.swyp.firsttodo.core.common.component.TaskCategoryList
+import com.swyp.firsttodo.core.common.component.TaskInputSection
 import com.swyp.firsttodo.core.common.util.HandleSideEffects
 import com.swyp.firsttodo.core.designsystem.theme.HaebomTheme
 import com.swyp.firsttodo.domain.model.habit.HabitDuration
-import com.swyp.firsttodo.presentation.common.component.HaebomLargeButton
-import com.swyp.firsttodo.presentation.common.component.HaebomMultiLineTextField
-import com.swyp.firsttodo.presentation.common.component.task.TaskCategoryList
-import com.swyp.firsttodo.presentation.common.component.task.TaskInputSection
 import com.swyp.firsttodo.presentation.habit.component.HabitDetailTopBar
 import com.swyp.firsttodo.presentation.habit.extension.displayName
 import com.swyp.firsttodo.presentation.main.snackbar.LocalSnackbarHostState
