+++ app/src/main/java/com/swyp/firsttodo/presentation/reward/list/RewardListContract.kt
@@ -4,9 +4,9 @@ import androidx.compose.runtime.Immutable
 import com.swyp.firsttodo.core.base.Async
 import com.swyp.firsttodo.core.base.UiEffect
 import com.swyp.firsttodo.core.base.UiState
+import com.swyp.firsttodo.core.common.type.HaebomHeaderTabType
 import com.swyp.firsttodo.domain.model.Role
 import com.swyp.firsttodo.domain.model.sticker.ChildStickerModel
-import com.swyp.firsttodo.presentation.common.component.HaebomHeaderTabType
 import com.swyp.firsttodo.presentation.reward.component.ChildRewardUiModel
 import com.swyp.firsttodo.presentation.reward.component.ParentRewardUiModel
 import com.swyp.firsttodo.presentation.reward.component.RewardFilterType
